requirements.txt
Flask==2.3.3
supabase==2.0.3
flask-talisman==1.0.0
python-dotenv==1.0.0
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.76
numpy==1.25.2
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.11.1
selenium==4.10.0
coverage==7.2.7